chore(dependabot): update svelte with other dependencies (#159)
Previously, svelte dependencies were updated independently because they
require to be updated to a new major version, and the configuration
required changes. The build was failing and prevented to update the
dependencies of other projects. So, updating svelte independently let
update the other dependencies. See
84b213d62102a7b9fd15cc3145cdb9d187d57fc7.

Svelte and sveltekit have been bumped to the latest available major
versions, so this is no longer needed to update them independently.